[fix] crud処理を分離
